*03* edit life and add item life

+++ f2/spw/GameReporter.java
@@ -1,6 +1,6 @@
 public interface GameReporter {
 
 	long getScore();
-	long getHpScore();
+	int getHpScore();
 
 }
